refactor: move NetworkRouterStatus into its own file
