+++ router/status.go
@@ -2,7 +2,6 @@ package router
 
 import (
 	"fmt"
-	"time"
 )
 
 type Status struct {
@@ -20,20 +19,6 @@ type Status struct {
 	Connections        []LocalConnectionStatus
 	Targets            []string
 	OverlayDiagnostics interface{}
-}
-
-type NetworkRouterStatus struct {
-	*Status
-	Interface    string
-	CaptureStats map[string]int
-	MACs         []MACStatus
-}
-
-type MACStatus struct {
-	Mac      string
-	Name     string
-	NickName string
-	LastSeen time.Time
 }
 
 type PeerStatus struct {
@@ -85,30 +70,6 @@ func NewStatus(router *Router) *Status {
 		NewLocalConnectionStatusSlice(router.ConnectionMaker),
 		NewTargetSlice(router.ConnectionMaker),
 		router.Overlay.Diagnostics()}
-}
-
-func NewNetworkRouterStatus(router *NetworkRouter) *NetworkRouterStatus {
-	return &NetworkRouterStatus{
-		NewStatus(router.Router),
-		router.Bridge.String(),
-		router.Bridge.Stats(),
-		NewMACStatusSlice(router.Macs)}
-}
-
-func NewMACStatusSlice(cache *MacCache) []MACStatus {
-	cache.RLock()
-	defer cache.RUnlock()
-
-	var slice []MACStatus
-	for key, entry := range cache.table {
-		slice = append(slice, MACStatus{
-			intmac(key).String(),
-			entry.peer.Name.String(),
-			entry.peer.NickName,
-			entry.lastSeen})
-	}
-
-	return slice
 }
 
 func NewPeerStatusSlice(peers *Peers) []PeerStatus {
